#18 コンポーネントの切り分け、整理
