+++ App.js
@@ -1,30 +1,12 @@
 import React from 'react';
-import { StyleSheet, Text, View, TouchableOpacity } from 'react-native';
-import GetImage from "./src/modules/GetImage"
-import { StatusBar } from 'expo-status-bar';
+import Home from "./src/modules/Home"
 import Header from "./src/modules/Header"
-
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
+
 const Stack = createNativeStackNavigator();
 
-export default function SplashScreen() {
-  const styles = StyleSheet.create({
-    container: {
-      flex: 1,
-      backgroundColor: '#fff',
-      alignItems: 'center',
-      justifyContent: 'center',
-    }
-  });
-  function HomeScreen() {
-    return (
-      <View style={styles.container}>
-        <GetImage />
-        <StatusBar style="auto" />
-      </View>
-    );
-  }
+const App = () => {
   return (
     <NavigationContainer>
       <Stack.Navigator
@@ -39,7 +21,7 @@ export default function SplashScreen() {
         }}>
         <Stack.Screen
           name="Home"
-          component={HomeScreen}
+          component={Home}
           // ここで、headerTitleにtitleコンポーネントを渡してあげます。
           options={{ headerTitle: Header }}
         />
@@ -48,3 +30,4 @@ export default function SplashScreen() {
   );
 }
 
+export default App;
